refactor(backend): tighten QueryBuilder, drop unused escape hatches
Route every caller through with_custom_filter() instead of mutating
.query directly, extract derived-field load options into their own
module, and remove with_explorer_rows_excluded/with_field_inclusion/
has_visibility/get_model_relationships now that nothing outside
QueryBuilder needs them.
