+++ apps/backend/src/rhesis/backend/app/utils/query_utils.py
@@ -1,9 +1,9 @@
 import logging
-from typing import Callable, Dict, List, Optional, Type, TypeVar
+from typing import Callable, List, Optional, Type, TypeVar
 from uuid import UUID
 
 from sqlalchemy import desc, inspect, or_
-from sqlalchemy.orm import Query, RelationshipProperty, Session, joinedload, selectinload
+from sqlalchemy.orm import Query, Session, joinedload, selectinload
 
 # Removed unused imports - legacy tenant functions no longer needed
 from rhesis.backend.app.utils.odata import apply_odata_filter
@@ -130,87 +130,13 @@ class QueryBuilder:
         return self
 
     def with_default_derived_field_loads(self, extra_chains: list | None = None) -> "QueryBuilder":
-        """Selectin-load comments/tasks/files/tags for this model, and for any
-        many-to-one/one-to-one relationship whose target model also has them.
-
-        Detail response schemas nest a related model's own derived fields
-        too (e.g. Test.prompt -> the nested PromptReference schema also gets
-        a "counts"/"tags" field, since Prompt has the same mixins), so those
-        need eager-loading as well -- not just this model's own. This is what
-        makes Test.prompt (near-1:1 with Test, so effectively N distinct
-        prompts per page) safe: without this, each row's prompt lazy-loads
-        its own comments/tasks/tags individually.
-
-        Checks the actual mixin class, not just attribute presence -- some
-        models (e.g. User.comments, "comments authored by this user") have an
-        unrelated attribute with the same name as a mixin's, which a plain
-        ``hasattr`` would wrongly match.
-
-        Safe to call unconditionally -- skips models/relationships that don't
-        have these mixins. Merges in any caller-supplied ``extra_chains`` too
-        (each a flat ``[name, ...]`` path, e.g. ``["_tags_relationship", "tag"]``),
-        deduped by full chain.
+        """Eager-load comments/tasks/files/tags for this model (and one level of
+        cascade into related models that carry the same mixins) -- see
+        ``derived_field_loads.derived_field_load_options`` for the policy.
         """
-        from rhesis.backend.app.models.mixins import (
-            CommentsMixin,
-            FilesMixin,
-            TagsMixin,
-            TasksMixin,
-        )
+        from rhesis.backend.app.utils.derived_field_loads import derived_field_load_options
 
-        derived_field_chains = (
-            (CommentsMixin, ("comments",)),
-            (TasksMixin, ("tasks",)),
-            (FilesMixin, ("files",)),
-            (TagsMixin, ("_tags_relationship", "tag")),
-        )
-
-        chains = list(extra_chains or [])
-        seen = {tuple(chain) for chain in chains}
-
-        def _add(chain: list) -> None:
-            key = tuple(chain)
-            if key not in seen:
-                seen.add(key)
-                chains.append(list(chain))
-
-        for mixin, chain in derived_field_chains:
-            if issubclass(self.model, mixin):
-                _add(list(chain))
-        for chain in chains:
-            # resolve_chain turns the runtime [name, ...] chain into a tuple of
-            # real attributes; include() picks joinedload vs. selectinload per hop
-            # from each one's own cardinality, whether the chain is a single hop
-            # (comments/tasks/files/tags) or multi-hop (_tags_relationship -> tag)
-            # -- no special-casing needed for either length.
-            self.with_related(include(*resolve_chain(self.model, chain)))
-
-        # Cascade one level into joined-in single-object relations (the ones
-        # with_related eager-loads via joinedload) whose
-        # target model also carries these mixins. These relationships are
-        # already strategy=joinedload (by convention, whether set by this
-        # call or by the caller) -- selectinload-ing the same attribute
-        # independently raises a loader-strategy conflict, so the nested load
-        # is chained off the existing joinedload instead of starting fresh.
-        seen_nested = set()
-        for rel_name, rel_prop in get_model_relationships(
-            self.model, skip_many_to_many=True, skip_one_to_many=True
-        ).items():
-            target_model = rel_prop.mapper.class_
-            for mixin, chain in derived_field_chains:
-                key = (rel_name, *chain)
-                if key in seen_nested or not issubclass(target_model, mixin):
-                    continue
-                seen_nested.add(key)
-                load = joinedload(getattr(self.model, rel_name))
-                nested_model = target_model
-                for nested_name in chain:
-                    load = load.selectinload(getattr(nested_model, nested_name))
-                    nested_rel_prop = inspect(nested_model).relationships.get(nested_name)
-                    if nested_rel_prop is not None:
-                        nested_model = nested_rel_prop.mapper.class_
-                self.query = self.query.options(load)
-        return self
+        return self.with_related(*derived_field_load_options(self.model, extra_chains))
 
     def _maybe_warn_load_count(self) -> None:
         if self._eager_load_count >= _MAX_EAGER_LOADS_WARN:
@@ -268,33 +194,22 @@ class QueryBuilder:
         Raises:
             ValueError: If organization_id is required but not provided
         """
-        if has_organization_id(self.model):
-            # Check if model is exempt from organization filtering
-            exempt_models = ["User", "Organization", "Token"]
-            if self.model.__name__ in exempt_models:
-                # For exempt models, apply organization filter if provided, but don't require it
-                if organization_id and (
-                    isinstance(organization_id, str)
-                    and organization_id.strip()
-                    or not isinstance(organization_id, str)
-                ):
-                    self.query = self.query.filter(self.model.organization_id == organization_id)
-            else:
-                # For non-exempt models, organization_id is required
-                if organization_id and (
-                    isinstance(organization_id, str)
-                    and organization_id.strip()
-                    or not isinstance(organization_id, str)
-                ):
-                    # Use direct organization_id filtering (optimized)
-                    self.query = self.query.filter(self.model.organization_id == organization_id)
-                else:
-                    # SECURITY: organization_id must be provided for models that have it
-                    raise ValueError(
-                        f"organization_id is required for {self.model.__name__} "
-                        "but was not provided. This is a security requirement to "
-                        "prevent data leakage across organizations."
-                    )
+        if not has_organization_id(self.model):
+            return self
+
+        provided = organization_id and (
+            not isinstance(organization_id, str) or organization_id.strip()
+        )
+        if provided:
+            self.query = self.query.filter(self.model.organization_id == organization_id)
+        elif self.model.__name__ not in ("User", "Organization", "Token"):
+            # SECURITY: organization_id must be provided for non-exempt models
+            # that have it, to prevent data leakage across organizations.
+            raise ValueError(
+                f"organization_id is required for {self.model.__name__} "
+                "but was not provided. This is a security requirement to "
+                "prevent data leakage across organizations."
+            )
         return self
 
     def with_project_filter(self, project_id: Optional[str] = None) -> "QueryBuilder":
@@ -416,29 +331,6 @@ class QueryBuilder:
         self.query = filter_func(self.query)
         return self
 
-    def with_explorer_rows_excluded(self) -> "QueryBuilder":
-        """Drop Explorer-owned rows -- they are listed through the /explorer API only."""
-        self.query = self.query.filter(self.model.explorer_row.is_(False))
-        return self
-
-    def with_field_inclusion(self, include_fields: str) -> "QueryBuilder":
-        """Apply field inclusion using SQLAlchemy undefer for deferred fields"""
-        if not include_fields:
-            return self
-
-        # Parse comma-separated fields
-        fields = [field.strip() for field in include_fields.split(",") if field.strip()]
-
-        # Apply undefer for each field that exists on the model
-        for field in fields:
-            if hasattr(self.model, field):
-                # Use SQLAlchemy's undefer to force load deferred fields
-                from sqlalchemy.orm import undefer
-
-                self.query = self.query.options(undefer(field))
-
-        return self
-
     def _apply_sorting(self):
         """Apply sorting if configured"""
         from rhesis.backend.app.utils.count_sort import (
@@ -537,53 +429,3 @@ def has_organization_id(model: Type[T]) -> bool:
 def has_project_id(model: Type[T]) -> bool:
     """Check if model has project_id column."""
     return hasattr(model, "project_id") or "project_id" in inspect(model).columns.keys()
-
-
-def has_visibility(model: Type[T]) -> bool:
-    """Check if model supports visibility filtering.
-
-    Requires a ``visibility`` column and an owner column (``user_id`` or
-    ``owner_user_id``).
-    """
-    columns = inspect(model).columns.keys()
-    has_owner = "user_id" in columns or "owner_user_id" in columns
-    return "visibility" in columns and has_owner
-
-
-def get_model_relationships(
-    model: Type, skip_many_to_many: bool = True, skip_one_to_many: bool = True
-) -> Dict[str, RelationshipProperty]:
-    """
-    Get relationships from a SQLAlchemy model.
-
-    Args:
-        model: The SQLAlchemy model class
-        skip_many_to_many: If True, excludes many-to-many relationships
-                          (those with secondary tables)
-        skip_one_to_many: If True, excludes one-to-many relationships (those with uselist=True)
-
-    Returns:
-        Dictionary of relationship name to RelationshipProperty
-    """
-    mapper = inspect(model)
-    relationships = {}
-
-    for rel in mapper.relationships:
-        # Use hierarchical filtering to avoid overlap between many-to-many and one-to-many
-
-        # First, check if it's many-to-many (has secondary table)
-        if getattr(rel, "secondary", None) is not None:
-            # This is a many-to-many relationship
-            if skip_many_to_many:
-                continue
-        # Then, check if it's one-to-many (uselist=True but no secondary table)
-        elif rel.uselist:
-            # This is a pure one-to-many relationship
-            if skip_one_to_many:
-                continue
-        # Otherwise, it's many-to-one or one-to-one (uselist=False, no secondary)
-
-        # Include this relationship
-        relationships[rel.key] = rel
-
-    return relationships
